WIP: Textbook functionals for Raviart-Thomas elements
Although the existing FIAT functionals are fine for tabulation and
hence normal finite element assembly, they under-integrate for the
purposes of interpolation. This results in sub-optimal convergence in
the Hdiv norm of the interpolation error, which should be

   ||u - I_q u||_H(div, T) \le C h^q |u|_H^{q+1}(T)

for RTq elements.

To fix this, replace the component evaluations in the interior with
proper integral moments and the normal component evaluations on the
faces with integral moments of v \cdot n.