restore old contains_point for sympy
In firedrake contains_point has sympy arguments given to it and I can't
think of a way to make it work with distance_to_point since you have to
actually do the computation to get the distance. Maybe there's a way to
force sympy to play nice? The error I get is
TypeError: cannot determine truth value of Relational