Attempt to unify MixedElement with the new UFL interface
Unfortunately this isn't quite there, as MixedElement has to deviate
from FiniteElementBase in a few places:
If a component is specified in either is_cellwise_constant() or
degree(), a domain must also be provided so the component can be
extracted using extract_component(). I've put this in as an optional
keyword-only argument, but I don't think this is the best approach.